chore: refactor `forge` for library use (#10117)
