+++ crates/forge/src/cmd/bind.rs
