+++ crates/forge/src/cmd/flatten.rs
