+++ crates/forge/src/cmd/remove.rs
